+++ R/NEWS.md
@@ -4,6 +4,14 @@
 
 * `geom_point2()` now accepts `size`-aesthetics when mapped to data.
 
+## New themes
+
+* `theme_light()`, a light and modern theme.
+
+## New plot-functions
+
+* Added plot-function for `performance::check_model()`.
+
 ## Bug fixes
 
 * Fixed issues with color codes in the flat-ui palette.